execute drainerfinisher also on delete events
